datastore: refactoring of relationship between worker and instance
- Split worker code out into a seperate module
- instance should now be easy to use without the worker
- Moved commit and uncommited_events out of DatastoreInstance
